refactor: rename network_search to network_utils
Rename network_search.dart to network_utils.dart to better reflect
its utility-like status. Update the corresponding tests, dashboard
tab import statements, and documentation file references.
